requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
flask_wtf==1.2.1
pytest==8.2.2
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
SQLAlchemy==1.4.46
